Specify the feature for creating a blog post
We start by writing out an acceptance test for the full feature we want to implement. In this case, we want to visit a blog post creation page, enter a title and body, save it, and then see on the results page that title and body, as well as confirming it’s in the database.

Red: No route matches [GET] "/blog_posts/new"

The first error we get is that there is no blog-posts/new route.